+++ tests/test_xrt.py
@@ -1,5 +1,5 @@
 import importlib
-import pynq.xrt
+import pynq._3rdparty.xrt
 import ctypes
 import os
 
@@ -17,7 +17,7 @@ class FakeXrt:
 def test_xrt_none(monkeypatch, recwarn):
     monkeypatch.delenv('XILINX_XRT', raising=False)
     monkeypatch.setattr(ctypes, 'CDLL', FakeXrt)
-    xrt = importlib.reload(pynq.xrt)
+    xrt = importlib.reload(pynq._3rdparty.xrt)
     assert xrt.XRT_SUPPORTED == False
     assert len(recwarn) == 0
 
@@ -26,7 +26,7 @@ def test_xrt_hw_emu(monkeypatch, recwarn):
     monkeypatch.setenv('XILINX_XRT', '/path/to/xrt')
     monkeypatch.setenv('XCL_EMULATION_MODE', 'hw_emu')
     monkeypatch.setattr(ctypes, 'CDLL', FakeXrt)
-    xrt = importlib.reload(pynq.xrt)
+    xrt = importlib.reload(pynq._3rdparty.xrt)
     assert xrt.XRT_SUPPORTED == True
     assert xrt.XRT_EMULATION == True
     assert xrt.libc.path == '/path/to/xrt/lib/libxrt_hwemu.so'
@@ -37,7 +37,7 @@ def test_xrt_sw_emu(monkeypatch, recwarn):
     monkeypatch.setenv('XILINX_XRT', '/path/to/xrt')
     monkeypatch.setenv('XCL_EMULATION_MODE', 'sw_emu')
     monkeypatch.setattr(ctypes, 'CDLL', FakeXrt)
-    xrt = importlib.reload(pynq.xrt)
+    xrt = importlib.reload(pynq._3rdparty.xrt)
     assert xrt.XRT_SUPPORTED == False
     assert len(recwarn) == 1
 
@@ -46,7 +46,7 @@ def test_xrt_wrong_emu(monkeypatch, recwarn):
     monkeypatch.setenv('XILINX_XRT', '/path/to/xrt')
     monkeypatch.setenv('XCL_EMULATION_MODE', 'wrong_emu')
     monkeypatch.setattr(ctypes, 'CDLL', FakeXrt)
-    xrt = importlib.reload(pynq.xrt)
+    xrt = importlib.reload(pynq._3rdparty.xrt)
     assert xrt.XRT_SUPPORTED == False
     assert len(recwarn) == 1
 
@@ -55,7 +55,7 @@ def test_xrt_normal(monkeypatch, recwarn):
     monkeypatch.setenv('XILINX_XRT', '/path/to/xrt')
     monkeypatch.delenv('XCL_EMULATION_MODE', raising=False)
     monkeypatch.setattr(ctypes, 'CDLL', FakeXrt)
-    xrt = importlib.reload(pynq.xrt)
+    xrt = importlib.reload(pynq._3rdparty.xrt)
     assert xrt.XRT_SUPPORTED == True
     assert xrt.XRT_EMULATION == False
     assert xrt.libc.path == '/path/to/xrt/lib/libxrt_core.so'
@@ -73,7 +73,7 @@ echo '{"runtime": {"build": {"version": "2.5.3"}}}'
     monkeypatch.delenv('XCL_EMULATION_MODE', raising=False)
     monkeypatch.setattr(ctypes, 'CDLL', FakeXrt)
     import pynq.pl_server.xrt_device
-    xrt = importlib.reload(pynq.xrt)
+    xrt = importlib.reload(pynq._3rdparty.xrt)
     xrt_device = importlib.reload(pynq.pl_server.xrt_device)
     assert xrt_device._xrt_version == (2, 5, 3)
 
@@ -89,7 +89,7 @@ exit 1
     monkeypatch.delenv('XCL_EMULATION_MODE', raising=False)
     monkeypatch.setattr(ctypes, 'CDLL', FakeXrt)
     import pynq.pl_server.xrt_device
-    xrt = importlib.reload(pynq.xrt)
+    xrt = importlib.reload(pynq._3rdparty.xrt)
     xrt_device = importlib.reload(pynq.pl_server.xrt_device)
     assert xrt_device._xrt_version == (0, 0, 0)
 
@@ -104,6 +104,6 @@ echo '{"runtime": {"build": {"version": "2.5.3"}}}'
     monkeypatch.delenv('XILINX_XRT', raising=False)
     monkeypatch.setattr(ctypes, 'CDLL', FakeXrt)
     import pynq.pl_server.xrt_device
-    xrt = importlib.reload(pynq.xrt)
+    xrt = importlib.reload(pynq._3rdparty.xrt)
     xrt_device = importlib.reload(pynq.pl_server.xrt_device)
     assert xrt_device._xrt_version == (0, 0, 0)
